cryptoauthlib: bump to version 3.7.6
Bump cryptoauthlib to the latest public release 3.7.6. For a list of
implemented features, bug fixes and API changes in this release, please
refer release notes below.Skip the QA check for the .so for this package
due to addition of package config files in this latest tag. Rename the
recipe to reflect the version info.

https://github.com/MicrochipTech/cryptoauthlib/blob/main/release_notes.md

Signed-off-by: Charan Pedumuru <charan.pedumuru@microchip.com>
Signed-off-by: Hari Prasath Gujulan Elango <hari.prasathge@microchip.com>
Signed-off-by: Dharma Balasubiramani <dharma.b@microchip.com>